remake login/logout screens into their own screens with proper DI and navigation
