+++ composeApp/src/commonMain/kotlin/com/jetbrains/kmpapp/App.kt
@@ -3,7 +3,7 @@ package com.jetbrains.kmpapp
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.navigator.Navigator
-import com.jetbrains.kmpapp.screens.LoginScreen
+import com.jetbrains.kmpapp.screens.auth.LoginScreen
 
 @Composable
 fun App() {
