+++ composeApp/src/commonMain/kotlin/com/jetbrains/kmpapp/screens/auth/AuthModel.kt
@@ -1,0 +1,6 @@
+package com.jetbrains.kmpapp.screens.auth
+
+import cafe.adriel.voyager.core.model.ScreenModel
+import dev.gitlive.firebase.auth.FirebaseAuth
+
+class AuthModel(val auth: FirebaseAuth) : ScreenModel
